Initial commit
A functional remote CLI client for the Faasten datastore. Supports
setting and getting database keys, as well as fetching and putting
blobs, by interacting with the REST frontend for Faasten.

Credentials are stored in ~/.config/fstn/credentials as TOML

Tested on Linux. Should work on other platforms, but YMMV for now.